news2md/requirements.txt
# news2md 依赖

# 基础
python-dotenv>=1.0.0
PyYAML>=6.0.1

# 爬虫
feedparser>=6.0.0           # Reddit RSS
requests>=2.31.0            # 3DM HTTP请求
beautifulsoup4>=4.12.0      # 3DM HTML解析

# AI 分析
google-genai>=0.3.0
